+++ ChangeLog.md
@@ -12,4 +12,5 @@ phpcpd 2.0.0
 * Added the `--git-repository` switch to calculate the software metrics for each revision of a Git repository.
 * Added `--names-exclude` switch to exclude filenames using glob patterns.
 * Removed the ability to count the Executable Lines of Code (ELOC) and the optional dependency on Bytekit.
+* The [Symfony Console](http://symfony.com/doc/current/components/console/) component is now used instead of the ConsoleTools component from the Zeta Components.
 * The [Version](http://github.com/sebastianbergmann/version) component is now used to manage the version number.